% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_weighted_criteria}
\alias{eval_weighted_criteria}
\title{Compute weighted quadratic errors statistics}
\usage{
eval_weighted_criteria(predictions, solutions, weights = NULL)
}
\arguments{
\item{predictions}{A vector of predicted values}

\item{solutions}{A vector of observed values}

\item{weights}{A vector of weights for the single observations (défault = \code{NULL}).
If \code{NULL}, then weights are taken as equal to 1}
}
\value{
The vector with the value of the MSE (index 1) and the R2 statistics
}
\description{
Given a vector of observed values \code{solutions} and a vector
of predicted values \code{predictions}, compute the Mean Squared Error (MSE) and
the percentage of explained variance (R2) statistics.
}
\examples{

}
\seealso{
\code{\link{SumModelGini}}
}
