% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{NormalizedGini}
\alias{NormalizedGini}
\title{Compute a Gini goodness of fit statistic}
\usage{
NormalizedGini(solutions, predictions, weights = NULL)
}
\arguments{
\item{solutions}{A vector of observed values}

\item{predictions}{A vector of predicted values}

\item{weights}{A vector of weights for the single observations (défault = \code{NULL}).
If \code{NULL}, then weights are taken as equal to 1}
}
\value{
The real number giving the value of the Gini index
}
\description{
Given a vector of observed values \code{solutions} and a vector
of predicted values \code{predictions}, the Gini index
measures how well the order of the predicted values corresponds
to the order of the observed values.
}
\details{
\itemize{
\item \code{solutions} is transformed into \code{rank(solutions)} before we compute
the gini coefficient
\item The weighted Gini index may not give meaningfull information if \code{weights}
depends on \code{predictions}
}
}
\examples{

set.seed(17)
x = runif(1000)
y = rnorm(mean = x, sd = 1, n = 1000)

NormalizedGini(solutions = x,
               predictions = y)

NormalizedGini(solutions = x,
               predictions = y,
               weights = x)

# be carefull with the weight argument
NormalizedGini(solutions = x,
               predictions = y,
               weights = abs(y))
# when the weights depends on \\code{predictions}, \\code{NormalizedGini}
# may gives strange results
}
\seealso{
\code{\link{SumModelGini}}
}
