% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlt_reg.R
\name{predict_rlt_reg}
\alias{predict_rlt_reg}
\title{Compute the prediction of a model built with \code{\link{rlt_reg}}}
\usage{
predict_rlt_reg(obj, newdata)
}
\arguments{
\item{obj}{A list output by \code{\link{rlt_reg}}}

\item{newdata}{A data.frame which contains the same variables as the ones
used for the training}
}
\value{
A list with the following elements :
\item{pred}{The vector of the predicted values for \code{phi}\eqn{(T')}
(with \eqn{T' = min(T, } \code{max_time}\eqn{)}) for the observations of \code{newdata}}
\item{surv}{The matrix which contains the estimated values of the survival curves at
\code{time_points}, for the observations of \code{newdata}}
\item{time_points}{The vector of the time points where the survival curves
are evaluated}
}
\description{
Given a model built wtih \code{\link{rlt_reg}},
\code{predict_rlt_reg} allows to get the predictions of the model for new
observations
}
\seealso{
\code{\link{rlt_reg}}
}
