% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrt_reg.R
\name{predict_rrt_reg}
\alias{predict_rrt_reg}
\title{Compute the prediction of a model built with \code{\link{rrt_reg}}}
\usage{
predict_rrt_reg(obj, newdata, type)
}
\arguments{
\item{obj}{A list output by \code{\link{rrt_reg}}}

\item{newdata}{A data.frame which contains the same variables as the ones
used for the training}

\item{type}{A character string which gives the type of terminal leaf estimator to use.
If \cite{"exp"}, IPCW used to grow the tree are used. If \cite{"KMloc"}, a Kaplan-Meier
estimator is fitted in each terminal leaf}
}
\value{
A list with the following elements :
\item{pred}{The vector of the predicted values for \code{phi}\eqn{(T')}
(with \eqn{T' = min(T, } \code{max_time}\eqn{)}) for the observations of \code{newdata}}
\item{surv}{The matrix which contains the estimated values of the survival curves at
\code{time_points}, for the observations of \code{newdata}}
\item{time_points}{The vector of the time points where the survival curves
are evaluated}
}
\description{
Given a model built wtih \code{\link{rrt_reg}},
\code{predict_rrt_reg} allows to get the predictions of the model for new
observations
}
\seealso{
\code{\link{rrt_reg}}
}
