% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSF_regression.R
\name{predict_RSF_regression}
\alias{predict_RSF_regression}
\title{Compute the prediction of a model built with \code{\link{RSF_regression}}}
\usage{
predict_RSF_regression(object, newdata)
}
\arguments{
\item{object}{A list output by \code{\link{RSF_regression}}}

\item{newdata}{A data.frame which contains the same variables as the ones
used for the training}
}
\value{
A list with the following elements :
\item{predicted}{The vector of the predicted values for \code{phi}\eqn{(T')}
(with \eqn{T' = min(T, } \code{max_time}\eqn{)}) for the observations of \code{newdata}}
\item{survival}{The matrix which contains the estimated values of the survival curves at
\code{time_points}, for the observations of \code{newdata}}
\item{time_points}{The vector of the time points where the survival curves
are evaluated}
}
\description{
Given a model built wtih \code{\link{RSF_regression}},
\code{predict_RSF_regression} allows to get the predictions of the model for new
observations
}
\examples{


}
\seealso{
\code{\link{RSF_regression}}
}
