% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showflow.R
\name{showflow}
\alias{showflow}
\title{showflow}
\usage{
showflow(dir = "./R")
}
\arguments{
\item{dir}{Directory/top-level of repository. Default is R sub-directory of working directory.}
}
\description{
Parse repository, document flows, and optionally show flows
Uses special code annotation to document flows
}
