% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{SumModelGini}
\alias{SumModelGini}
\title{helper function}
\usage{
SumModelGini(solutions, predictions, weights)
}
\description{
function called by \code{\link{NormalizedGini}}
}
\seealso{
\code{\link{NormalizedGini}}
}
