% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sword.R
\name{sword}
\alias{sword}
\title{sword}
\usage{
sword(repo = ".", dir = "R", flows = NULL)
}
\arguments{
\item{repo}{Directory/top-level of repository. Default is working directory.}

\item{dir}{Sub-directory of R scripts. Default is R sub-directory of working directory.}

\item{flows}{Optional, which flows to show}
}
\description{
Parse repository, document flows, and show DAGs
Uses special code annotation to document flows
}
