% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_regression.R
\name{predict_Cox_regression}
\alias{predict_Cox_regression}
\title{Compute the prediction of a model built with Cox_regression}
\usage{
predict_Cox_regression(object, newdata)
}
\arguments{
\item{object}{An output from Cox_regression}

\item{newdata}{A dataframe which contains the same variables as the ones used for the training}
}
\value{
A list with the following elements :
\item{predicted}{A vector of the predicted values for phi(T)}
\item{survival}{A dataframe (matrix) of the predicted survival curves given by the Cox model}
\item{time_points}{A vector of the time points where the survival curves are evaluated}
}
\description{
Compute the prediction of a model built with Cox_regression
}
\examples{


}
