% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showflow.R
\name{showflow}
\alias{showflow}
\title{showflow}
\usage{
showflow(repo = ".", dir = "R", flows = NULL)
}
\arguments{
\item{repo}{Directory/top-level of repository. Default is working directory.}

\item{dir}{Sub-directory of R scripts. Default is R sub-directory of working directory.}

\item{flows}{Optional, which flows to show}
}
\description{
Parse repository, document flows, and show flows
Uses special code annotation to document flows
}
